monitor-bot/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
psutil>=5.9.0
py-cpuinfo>=9.0.0
python-dotenv>=1.0.0
configparser>=5.3.0
asyncio>=3.4.3
